code/src/Mock/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
python-engineio==4.4.0
python-socketio==5.8.0
uvicorn==0.21.1
yarl==1.8.2
